Minor formatting tweaks on the `help` command (#22)
* Minor formatting tweaks on the `help` command

There were a few commas and capitalizations off. Also put some other stuff in italics or quotes for readibility.

* More formatting tweaks; error msgs slightly off

Some more formatting tweaks, for readibility.

Also fixed a minor typo on `add`: "atleast" -> "at least"

`check` and `start` also appeared to have a minor bug in an error message, as it displayed `stop: ` instead of the respective command. Read this as a copy-paste of a chunk of code and then forgetting to update it to the relevant command, but if I misunderstood and the behavior is actually intentional and I just missed the point, just ignore that bit :)